backend/requirements.txt
fastapi
uvicorn[standard]
weaviate-client==4.*
pypdf
python-multipart
anthropic
python-dotenv
